Field Of View and Additions to Entity State Machine
This adds the Field of View for Entities that could be used for when the enemies are patrolling. This is one of my main steps to getting the NPCs to get into battle and chase the player.

Very important for the State Machine since I want to allow the NPCs to see the player, but not through walls. This will also help me change the states so that they can actually chase the player.