The Start for Items, Status Effects, Entities
I've begun the process for getting the UI running with the inventory system with the different types of Items. I'm pre-emptively posting this in the case that something may go wrong later.

The Entity has a few more features to make it easier to add temporary effects onto.

The Inventory System hasn't been worked onto yet, but it'll be up next now that I'm done with added the different types of items I want.

Items were actively worked on during this commit. In theory items could be added in now with what's there. They are available as Scriptable Objects.